requirements/common.txt
dj-rest-auth==2.2.5

Django==4.0.3
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==21.1
django-google-tag-manager==0.0.5
django-import-export==2.8.0
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-quill-editor==0.1.40
django-storages==1.12.3
django-tinymce==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-nested-routers==0.93.4
drf-spectacular==0.23.1
python-decouple==3.6
requests==2.27.1
